feat: move CommunityFeedFilter and CommunityModerationFeedFilter to commons
Move both community feed filters to commons/ui/feeds/ with IAccount and
ICacheProvider interfaces. Add mapFlattenNotesIntoSet and filterNotesIntoSet
to ICacheProvider for cache iteration. Leave typealiases at original locations.
Update ViewModels to pass LocalCache as ICacheProvider.
